Add files via upload
Thanks to a tutorial on freeCodeCamp.org's YouTube channel, I was able to make a memory game and learn about various methods, such as createElement() and querySelector(), as well as how to amend those elements using setAttribute() and appendChild(). The one modification I made to the original tutorial was adding a means of resetting the game after the game was over. There is probably a more elegant way of resetting the state of things, but I merely used an event listener when a reset button was clicked that caused the location.reload() to happen